+++ mycodo/outputs/mqtt.py
@@ -177,6 +177,31 @@ OUTPUT_INFORMATION = {
             'required': True,
             'name': lazy_gettext('Current (Amps)'),
             'phrase': lazy_gettext('The current draw of the device being controlled')
+        },
+        {
+            'id': 'login',
+            'type': 'bool',
+            'default_value': False,
+            'name': 'Use Login',
+            'phrase': 'Send login credentials'
+        },
+        {
+            'id': 'username',
+            'type': 'text',
+            'default_value': 'user',
+            'required': False,
+            'name': lazy_gettext('Username'),
+            'phrase': lazy_gettext('Username for connecting to the server')
+        },
+        {
+            'id': 'password',
+            'type': 'text',
+            'default_value': '',
+            'required': False,
+            'name': lazy_gettext('Password'),
+            'phrase': "{} {}".format(
+                lazy_gettext('Password for connecting to the server.'),
+                lazy_gettext('Leave blank to disable.'))
         }
     ]
 }
@@ -211,6 +236,15 @@ class OutputModule(AbstractOutput):
 
     def output_switch(self, state, output_type=None, amount=None, output_channel=0):
         try:
+            auth_dict = None
+            if self.options_channels['login'][0]:
+                if not self.options_channels['password'][0]:
+                    self.options_channels['password'][0] = None
+                auth_dict = {
+                    "username": self.options_channels['username'][0],
+                    "password": self.options_channels['password'][0]
+                }
+
             if state == 'on':
                 self.publish.single(
                     self.options_channels['topic'][0],
@@ -218,7 +252,8 @@ class OutputModule(AbstractOutput):
                     hostname=self.options_channels['hostname'][0],
                     port=self.options_channels['port'][0],
                     client_id=self.options_channels['clientid'][0],
-                    keepalive=self.options_channels['keepalive'][0])
+                    keepalive=self.options_channels['keepalive'][0],
+                    auth=auth_dict)
                 self.output_states[output_channel] = True
             elif state == 'off':
                 self.publish.single(
@@ -227,7 +262,8 @@ class OutputModule(AbstractOutput):
                     hostname=self.options_channels['hostname'][0],
                     port=self.options_channels['port'][0],
                     client_id=self.options_channels['clientid'][0],
-                    keepalive=self.options_channels['keepalive'][0])
+                    keepalive=self.options_channels['keepalive'][0],
+                    auth=auth_dict)
                 self.output_states[output_channel] = False
         except Exception as e:
             self.logger.error("State change error: {}".format(e))